requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
numpy
tqdm
